refactor(repo C9): extract collection helpers to engine/repository/collections.ts
Moves getCollectionByName, listCollections, removeCollection,
renameCollection, getAllCollections.

Build: clean. Tests: 970/970 pass.
