+++ engine/repository/index.ts
@@ -12,3 +12,4 @@ export * from "./docid.js";
 export * from "./rerank-queue.js";
 export * from "./chunking.js";
 export * from "./fts.js";
+export * from "./collections.js";
